Move extract_auth_from_url to non-Django utils.
