+++ go/base/utils.py
@@ -4,7 +4,6 @@ import csv
 import codecs
 from decimal import Decimal, ROUND_DOWN
 from StringIO import StringIO
-from urlparse import urlparse, urlunparse
 
 from django.http import Http404, HttpResponse
 from django.conf import settings
@@ -212,23 +211,6 @@ def get_router_view_definition(router_type, router=None):
     if not hasattr(router_pkg, 'view_definition'):
         return RouterViewDefinitionBase(router_def)
     return router_pkg.view_definition.RouterViewDefinition(router_def)
-
-
-def extract_auth_from_url(url):
-    parse_result = urlparse(url)
-    if parse_result.username:
-        auth = (parse_result.username, parse_result.password)
-        url = urlunparse(
-            (parse_result.scheme,
-             ('%s:%s' % (parse_result.hostname, parse_result.port)
-              if parse_result.port
-              else parse_result.hostname),
-             parse_result.path,
-             parse_result.params,
-             parse_result.query,
-             parse_result.fragment))
-        return auth, url
-    return None, url
 
 
 def format_currency(
